+++ examples/server/server.c
@@ -36,15 +36,11 @@
 #endif
 
 #ifdef WOLFSSL_WOLFSENTRY_HOOKS
-#    include <wolfsentry/wolfsentry.h>
-#    include <wolfsentry/wolfsentry_util.h>
-#    include <wolfsentry/wolfsentry_json.h>
-
+#include <wolfsentry/wolfsentry.h>
 #if !defined(NO_FILESYSTEM) && !defined(WOLFSENTRY_NO_JSON)
 static const char *wolfsentry_config_path = NULL;
 #endif
-
-#endif
+#endif /* WOLFSSL_WOLFSENTRY_HOOKS */
 
 #if defined(WOLFSSL_MDK_ARM) || defined(WOLFSSL_KEIL_TCP_NET)
         #include <stdio.h>
@@ -1119,7 +1115,6 @@ THREAD_RETURN WOLFSSL_THREAD server_test(void* args)
     SSL_CTX*    ctx    = 0;
     SSL*        ssl    = 0;
 #ifdef WOLFSSL_WOLFSENTRY_HOOKS
-    struct wolfsentry_context *wolfsentry = NULL;
     wolfsentry_errcode_t wolfsentry_ret;
 #endif
 
@@ -1984,125 +1979,8 @@ THREAD_RETURN WOLFSSL_THREAD server_test(void* args)
         err_sys_ex(catastrophic, "unable to get ctx");
 
 #ifdef WOLFSSL_WOLFSENTRY_HOOKS
-    wolfsentry_ret =  wolfsentry_init(NULL /* hpi */, NULL /* default config */,
-                                      &wolfsentry);
-    if (wolfsentry_ret < 0) {
-        fprintf(stderr, "wolfsentry_init() returned " WOLFSENTRY_ERROR_FMT "\n",
-                WOLFSENTRY_ERROR_FMT_ARGS(wolfsentry_ret));
-        err_sys_ex(catastrophic, "unable to initialize wolfSentry");
-    }
-
-    if (wolfsentry_data_index < 0)
-        wolfsentry_data_index = wolfSSL_get_ex_new_index(0, NULL, NULL, NULL,
-                                                         NULL);
-
-#if !defined(NO_FILESYSTEM) && !defined(WOLFSENTRY_NO_JSON)
-    if (wolfsentry_config_path != NULL) {
-        char buf[512], err_buf[512];
-        struct wolfsentry_json_process_state *jps;
-
-        FILE *f = fopen(wolfsentry_config_path, "r");
-
-        if (f == NULL) {
-            fprintf(stderr, "fopen(%s): %s\n",wolfsentry_config_path,strerror(errno));
-            err_sys_ex(catastrophic, "unable to open wolfSentry config file");
-        }
-
-        if ((wolfsentry_ret = wolfsentry_config_json_init(
-                 wolfsentry,
-                 WOLFSENTRY_CONFIG_LOAD_FLAG_NONE,
-                 &jps)) < 0) {
-            fprintf(stderr, "wolfsentry_config_json_init() returned "
-                    WOLFSENTRY_ERROR_FMT "\n",
-                    WOLFSENTRY_ERROR_FMT_ARGS(wolfsentry_ret));
-            err_sys_ex(catastrophic, "error while initlalizing wolfSentry config parser");
-        }
-
-        for (;;) {
-            size_t n = fread(buf, 1, sizeof buf, f);
-            if ((n < sizeof buf) && ferror(f)) {
-                fprintf(stderr,"fread(%s): %s\n",wolfsentry_config_path, strerror(errno));
-                err_sys_ex(catastrophic, "error while reading wolfSentry config file");
-            }
-
-            wolfsentry_ret = wolfsentry_config_json_feed(jps, buf, n, err_buf, sizeof err_buf);
-            if (wolfsentry_ret < 0) {
-                fprintf(stderr, "%.*s\n", (int)sizeof err_buf, err_buf);
-                err_sys_ex(catastrophic, "error while loading wolfSentry config file");
-            }
-            if ((n < sizeof buf) && feof(f))
-                break;
-        }
-        fclose(f);
-
-        if ((wolfsentry_ret = wolfsentry_config_json_fini(jps, err_buf, sizeof err_buf)) < 0) {
-            fprintf(stderr, "%.*s\n", (int)sizeof err_buf, err_buf);
-            err_sys_ex(catastrophic, "error while loading wolfSentry config file");
-        }
-
-    } else
-#endif /* !defined(NO_FILESYSTEM) && !defined(WOLFSENTRY_NO_JSON) */
-    {
-        struct wolfsentry_route_table *table;
-
-        if ((wolfsentry_ret = wolfsentry_route_get_table_static(wolfsentry,
-                                                                &table)) < 0)
-            fprintf(stderr, "wolfsentry_route_get_table_static() returned "
-                    WOLFSENTRY_ERROR_FMT "\n",
-                    WOLFSENTRY_ERROR_FMT_ARGS(wolfsentry_ret));
-        if (wolfsentry_ret >= 0) {
-            if ((wolfsentry_ret = wolfsentry_route_table_default_policy_set(
-                     wolfsentry, table,
-                     WOLFSENTRY_ACTION_RES_REJECT|WOLFSENTRY_ACTION_RES_STOP))
-                < 0)
-                fprintf(stderr,
-                        "wolfsentry_route_table_default_policy_set() returned "
-                        WOLFSENTRY_ERROR_FMT "\n",
-                        WOLFSENTRY_ERROR_FMT_ARGS(wolfsentry_ret));
-        }
-
-        if (wolfsentry_ret >= 0) {
-            struct {
-                struct wolfsentry_sockaddr sa;
-                byte buf[16];
-            } remote, local;
-            wolfsentry_ent_id_t id;
-            wolfsentry_action_res_t action_results;
-
-            memset(&remote, 0, sizeof remote);
-            memset(&local, 0, sizeof local);
-#ifdef TEST_IPV6
-            remote.sa.sa_family = local.sa.sa_family = AF_INET6;
-            remote.sa.addr_len = 128;
-            memcpy(remote.sa.addr, "\000\000\000\000\000\000\000\000\000\000\000\000\000\000\000\001", 16);
-#else
-            remote.sa.sa_family = local.sa.sa_family = AF_INET;
-            remote.sa.addr_len = 32;
-            memcpy(remote.sa.addr, "\177\000\000\001", 4);
-#endif
-
-            if ((wolfsentry_ret = wolfsentry_route_insert_static
-                 (wolfsentry, NULL /* caller_context */, &remote.sa, &local.sa,
-                  WOLFSENTRY_ROUTE_FLAG_GREENLISTED              |
-                  WOLFSENTRY_ROUTE_FLAG_DIRECTION_IN             |
-                  WOLFSENTRY_ROUTE_FLAG_PARENT_EVENT_WILDCARD    |
-                  WOLFSENTRY_ROUTE_FLAG_REMOTE_INTERFACE_WILDCARD|
-                  WOLFSENTRY_ROUTE_FLAG_LOCAL_INTERFACE_WILDCARD |
-                  WOLFSENTRY_ROUTE_FLAG_SA_LOCAL_ADDR_WILDCARD   |
-                  WOLFSENTRY_ROUTE_FLAG_SA_PROTO_WILDCARD        |
-                  WOLFSENTRY_ROUTE_FLAG_SA_REMOTE_PORT_WILDCARD  |
-                  WOLFSENTRY_ROUTE_FLAG_SA_LOCAL_PORT_WILDCARD,
-                  0 /* event_label_len */, 0 /* event_label */, &id,
-                  &action_results)) < 0)
-                fprintf(stderr, "wolfsentry_route_insert_static() returned "
-                        WOLFSENTRY_ERROR_FMT "\n",
-                        WOLFSENTRY_ERROR_FMT_ARGS(wolfsentry_ret));
-        }
-
-        if (wolfsentry_ret < 0)
-            err_sys_ex(catastrophic, "unable to configure route table");
-    }
-
+    if (wolfsentry_setup(&wolfsentry, wolfsentry_config_path, WOLFSENTRY_ROUTE_FLAG_DIRECTION_IN) < 0)
+        err_sys("unable to initialize wolfSentry");
 
     if (wolfSSL_CTX_set_AcceptFilter(
             ctx,
